requirements.txt
mujoco
dm-control
numpy<2.0.0
gymnasium
pyyaml
imageio
